[refactor] : LottoComparisonManager와 LottoPrizeManager 클래스 분리
