+++ src/step1-index.js
@@ -5,7 +5,7 @@
 
 import readLineAsync from "./readLineAsync.js";
 import UserLottos from "./UserLottos.js";
-import LottoManager from "./LottoManager.js";
+import LottoComparisonManager from "./LottoComparisonManager.js";
 import { printUserLottos, printWinningResult, printROI } from "./output.js";
 import {
   inputAskForRestart,
@@ -13,32 +13,38 @@ import {
   inputPrice,
   inputWinningNumbers,
 } from "./input.js";
+import LottoPrizeManager from "./LottoPrizeManager.js";
 
 async function run() {
-  while (true) {
-    const price = await inputPrice();
-    const userLottos = new UserLottos(price);
-    printUserLottos(userLottos);
+  const price = await inputPrice();
+  const userLottos = new UserLottos(price);
+  printUserLottos(userLottos);
 
-    const winningNumbers = await inputWinningNumbers();
-    const bonusNumber = await inputBonusNumber(winningNumbers);
+  const winningNumbers = await inputWinningNumbers();
+  const bonusNumber = await inputBonusNumber(winningNumbers);
 
-    const lottoManager = new LottoManager(
-      userLottos.lottos,
-      winningNumbers,
-      bonusNumber
-    );
+  const lottoComparisonManager = new LottoComparisonManager(
+    userLottos.lottos,
+    winningNumbers,
+    bonusNumber
+  );
 
-    const countResults = lottoManager.countMatchingNumbers();
-    const isContainBonusNumbers =
-      lottoManager.containsBonusNumbers(countResults);
-    lottoManager.calculateWinnings(countResults, isContainBonusNumbers);
+  const countResults = lottoComparisonManager.countMatchingNumbers();
+  const isContainBonusNumbers =
+    lottoComparisonManager.containsBonusNumbers(countResults);
+  console.log(countResults, isContainBonusNumbers);
+  const lottoPrizeManager = new LottoPrizeManager(
+    countResults,
+    isContainBonusNumbers
+  );
 
-    printWinningResult(lottoManager.prizeResult);
-    printROI(lottoManager.calculateROI(price));
+  lottoPrizeManager.calculateWinnings();
+  printWinningResult(lottoPrizeManager.prizeResult);
+  printROI(lottoPrizeManager.calculateROI(price));
 
-    const isRestart = await inputAskForRestart();
-    if (isRestart === "n") break;
+  const isRestart = await inputAskForRestart();
+  if (isRestart === "n") {
+    return await run();
   }
 }
 run();
